(PC-21171)[PRO] refactor: clean rff fields
