+++ pro/src/ui-kit/form_rff/fields/CheckboxField/__specs__/CheckboxField.spec.jsx
