--- pro/src/ui-kit/form_rff/fields/SiretField/TooltipInvalidSiret/index.ts
@@ -1,1 +1,0 @@
-export { default as TooltipInvalidSiret } from './TooltipInvalidSiret'
