--- pro/src/ui-kit/form_rff/fields/SiretField/TooltipValidSiret/index.ts
@@ -1,1 +1,0 @@
-export { default as TooltipValidSiret } from './TooltipValidSiret'
