--- pro/src/ui-kit/form_rff/fields/SiretField/index.ts
@@ -1,1 +1,0 @@
-export { default as SiretField } from './SiretField'
