+++ pro/src/ui-kit/index.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
-export { default as CheckboxField } from './CheckboxField'
+export { default as CheckboxField } from './form_rff/fields/CheckboxField'
 export { default as Divider } from './Divider'
 export { default as SubmitButton } from './SubmitButton'
 export { default as Tag } from './Tag'
